+++ src/components/Router/Router.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Home from '../../pages/Home';
-import QuizWelcome from '../../pages/QuizWelcome';
-import Quiz from '../../pages/Quiz';
+import MeetGuy from '../../pages/MeetGuy';
+import Quiz from '../../pages/Questionaire';
 import ROUTES from '../Router/RouteConfig';
 
 // To Do - How do we unit test this.
@@ -12,7 +12,7 @@ const Router = () => {
     <BrowserRouter>
       <Switch>
         <Route exact path={ROUTES.ROUTE_HOME} render={() => <Home />} />
-        <Route path={ROUTES.ROUTE_QUIZHOME} render={() => <QuizWelcome />} />
+        <Route path={ROUTES.ROUTE_QUIZHOME} render={() => <MeetGuy />} />
         <Route path={ROUTES.ROUTE_QUIZ} render={() => <Quiz />} />
       </Switch>
     </BrowserRouter>
